requirements.txt
Pillow
pillow-heif
pillow-avif-plugin
